[SCFToCalyx] Add support for sequential read MemRefType memories
If a FuncOp argument is of MemRefType and has the boolean attribute
calyx.sequential_reads set to true, consider the read acceses to that
memory as being sequential, which adds a "read_en" signal to the
external memory interface. This "read_en" signal is driven by a
memref::LoadOp to a memref with has sequential reads.

A register that stores the data read from memory and being forwarded
to the next group is created.

Currently, the memory access ports contain a single "done" signal,
therefore it's not possible to perfom a memory write and load at the same time.

This allows interfacing (with extra logic) with protocols such as
AXI4.

Signed-off-by: Sergi Granell <xerpi.g.12@gmail.com>